Temporary fix for dropdown field font not rendering correctly
This will be permanently fixed when the UI is rewritten. The fieldfont and the spriteBatch are rendered deferred instead of with depth. Adding depth will fix this, however SpriteBatch has some strange behavior regarding depth that doesn't play nicely with the field font renderer. The UI rework fixes this by using a custom sprite batch in the RenderSystem instead, allowing for a custom projection matrix that can be shared with the field font renderer.

The current fix cannot handle dropdowns within dropdowns. The UI rework will handle that case. This isn't a problem currently (and probably won't be) since nested dropdowns aren't used in the game.